Refactoring to allow for one config instead of needing three.
This PR refactors the pipeline to remove redundancy in how configs are used. The goal with the changes here are to allow for ease of use while still allowing for flexibility. Overall, only one config will need to be supplied now and each command line tool can be specified in the command line argument.